+++ contrib/python/tests/python/pants_test/contrib/python/checks/tasks/checkstyle/BUILD
@@ -15,6 +15,7 @@ python_tests(
     'src/python/pants/backend/python/tasks',
     'src/python/pants/base:build_environment',
     'src/python/pants/base:exceptions',
+    'src/python/pants/python',
     'src/python/pants/util:contextutil',
     'src/python/pants/util:dirutil',
     'tests/python/pants_test/backend/python/tasks:python_task_test_base',
