+++ src/python/pants/backend/python/subsystems/BUILD
@@ -14,6 +14,7 @@ python_library(
     'src/python/pants/base:exceptions',
     'src/python/pants/build_graph',
     'src/python/pants/option',
+    'src/python/pants/python',
     'src/python/pants/subsystem',
     'src/python/pants/util:memo',
   ],
