+++ src/python/pants/init/BUILD
@@ -9,7 +9,6 @@ python_library(
     '3rdparty/python:setuptools',
     '3rdparty/python:wheel',
     '3rdparty/python/twitter/commons:twitter.common.collections',
-    'src/python/pants/backend/python/subsystems',
     'src/python/pants/base:build_environment',
     'src/python/pants/base:build_root',
     'src/python/pants/base:exception_sink',
@@ -34,6 +33,7 @@ python_library(
     'src/python/pants/goal:run_tracker',
     'src/python/pants/option',
     'src/python/pants/process',
+    'src/python/pants/python',
     'src/python/pants/rules/core',
     'src/python/pants/subsystem',
     'src/python/pants/util:dirutil',
