+++ src/python/pants/python/BUILD
@@ -1,0 +1,11 @@
+# Copyright 2020 Pants project contributors (see CONTRIBUTORS.md).
+# Licensed under the Apache License, Version 2.0 (see LICENSE).
+
+
+python_library(
+  dependencies=[
+    '3rdparty/python:pex',
+    'src/python/pants/subsystem',
+    'src/python/pants/util:memo',
+  ]
+)
