+++ tests/python/pants_test/backend/codegen/thrift/python/BUILD
@@ -7,9 +7,9 @@ python_tests(
     'src/python/pants/backend/codegen/thrift/lib',
     'src/python/pants/backend/codegen/thrift/python',
     'src/python/pants/backend/python:interpreter_cache',
-    'src/python/pants/backend/python/subsystems',
     'src/python/pants/backend/python/targets',
     'src/python/pants/base:build_environment',
+    'src/python/pants/python',
     'src/python/pants/util:contextutil',
     'src/python/pants/testutil/subsystem',
     'src/python/pants/testutil:task_test_base',
