+++ tests/python/pants_test/backend/codegen/thrift/python/test_apache_thrift_py_gen.py
@@ -12,9 +12,9 @@ from pants.backend.codegen.thrift.lib.thrift import Thrift
 from pants.backend.codegen.thrift.python.apache_thrift_py_gen import ApacheThriftPyGen
 from pants.backend.codegen.thrift.python.python_thrift_library import PythonThriftLibrary
 from pants.backend.python.interpreter_cache import PythonInterpreterCache
-from pants.backend.python.subsystems.python_repos import PythonRepos
 from pants.backend.python.targets.python_library import PythonLibrary
 from pants.base.build_environment import get_buildroot
+from pants.python.python_repos import PythonRepos
 from pants.testutil.subsystem.util import global_subsystem_instance
 from pants.testutil.task_test_base import TaskTestBase
