+++ tests/python/pants_test/backend/python/tasks/util/build_local_dists_test_base.py
@@ -8,12 +8,12 @@ from enum import Enum
 from pants.backend.native.register import rules as native_backend_rules
 from pants.backend.native.subsystems.libc_dev import LibcDev
 from pants.backend.native.subsystems.native_build_step import ToolchainVariant
-from pants.backend.python.subsystems.python_repos import PythonRepos
 from pants.backend.python.tasks.build_local_python_distributions import (
   BuildLocalPythonDistributions,
 )
 from pants.backend.python.tasks.resolve_requirements import ResolveRequirements
 from pants.backend.python.tasks.select_interpreter import SelectInterpreter
+from pants.python.python_repos import PythonRepos
 from pants.testutil.task_test_base import DeclarativeTaskTestMixin
 from pants.util.collections import assert_single_element
 from pants.util.enums import match
